Moved utility files to utils directory
